Changed the turtle water

+++ MRIPIDM/helperFuncs.h
@@ -3,7 +3,7 @@
 
  float* ippsMalloc_32fHF(int len) ;
  void ippsFreeHF(float* ptr);
- void ippsZero_64fHF(float* ptr, int len);
+ void ippsZero_64fHF(double* ptr, int len);
  void ippsSum_32fHF(const float* src, int len, float* sum);
 
 #endif
